pyproject.toml
[tool.poetry]
name = "djangomailing"
version = "0.1.0"
description = ""
authors = ["ryfe <denis.koptelev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
django-apscheduler = "^0.6.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
python-jose = "^3.3.0"
django-crispy-forms = "^2.1"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
